API Revision: losing details added to Winning API

+++ database/migrations/2020_12_03_213935_create_user_achievements_table.php
@@ -11,8 +11,8 @@ class CreateUserAchievementsTable extends Migration
         Schema::create('user_achievements', static function (Blueprint $table) {
             $table->id();
             $table->string('user_id');
-            $table->unsignedSmallInteger('coins_change');
-            $table->unsignedSmallInteger('trophies_change');
+            $table->integer('coins_change');
+            $table->integer('trophies_change');
             $table->timestamp('created_at');
         });
